requirements.txt
pyTelegramBotAPI~=4.26.0
python-dotenv~=1.0.1
